Remove the CleanupSubscriber interface/implementation with its adding/removing logic, in favour of a simple subscriber that closes cursors.
