--- app/src/main/java/org/projectbuendia/client/events/CleanupSubscriber.java
@@ -1,24 +1,0 @@
-// Copyright 2015 The Project Buendia Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not
-// use this file except in compliance with the License.  You may obtain a copy
-// of the License at: http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software distrib-
-// uted under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
-// OR CONDITIONS OF ANY KIND, either express or implied.  See the License for
-// specific language governing permissions and limitations under the License.
-
-package org.projectbuendia.client.events;
-
-import de.greenrobot.event.NoSubscriberEvent;
-
-/** A {@link CrudEventBus} subscriber that gets invoked when all subscribers have been unregistered. */
-public interface CleanupSubscriber {
-
-    /** Called when an event is posted but no subscribers are registered for that event. */
-    void onEvent(NoSubscriberEvent event);
-
-    /** Called when all subscribers on a {@link CrudEventBus} have been unregistered. */
-    void onAllUnregistered();
-}
