+++ app/src/main/java/org/projectbuendia/client/events/CrudEventBus.java
@@ -19,23 +19,9 @@ package org.projectbuendia.client.events;
  * if an event is posted for which there are no listeners.
  */
 public interface CrudEventBus {
-
     void register(Object subscriber);
 
     void unregister(Object subscriber);
 
     void post(Object event);
-
-    /**
-     * Registers a {@link CleanupSubscriber} that gets invoked when all subscribers have been
-     * unregistered.
-     * @param subscriber the subscriber to invoke or {@code null} to disable the callback
-     */
-    void registerCleanupSubscriber(CleanupSubscriber subscriber);
-
-    /**
-     * Unregistered a {@link CleanupSubscriber} that was previously registered with
-     * {@link #registerCleanupSubscriber}.
-     */
-    void unregisterCleanupSubscriber(CleanupSubscriber subscriber);
 }
