+++ app/src/main/java/org/projectbuendia/client/events/DefaultCrudEventBus.java
@@ -11,83 +11,42 @@
 
 package org.projectbuendia.client.events;
 
-import java.util.HashSet;
-import java.util.Set;
+import org.projectbuendia.client.events.data.AppLocationTreeFetchedEvent;
+import org.projectbuendia.client.events.data.TypedCursorFetchedEvent;
 
 import de.greenrobot.event.EventBus;
+import de.greenrobot.event.NoSubscriberEvent;
 
 /** An {@link EventBus} wrapper to be used by asynchronous CRUD operations on the data model. */
 public final class DefaultCrudEventBus implements CrudEventBus {
+    private final EventBus mBus;
 
-    private final Object mSubscribersLock = new Object();
-
-    private final EventBus mWrapped;
-    private final Set<Object> mSubscribers;
-    private CleanupSubscriber mCleanupSubscriber;
+    /** Creates a new {@link DefaultCrudEventBus} that wraps the specified {@link EventBus}. */
+    DefaultCrudEventBus(EventBus bus) {
+        mBus = bus;
+        bus.register(this);
+    }
 
     @Override public void register(Object subscriber) {
-        if (subscriber instanceof CleanupSubscriber) {
-            throw new IllegalArgumentException(
-                "CleanupSubscribers must be registered with registerCleanupSubscriber().");
-        }
-
-        mWrapped.register(subscriber);
-
-        synchronized (mSubscribersLock) {
-            mSubscribers.add(subscriber);
-        }
+        mBus.register(subscriber);
     }
 
     @Override public void unregister(Object subscriber) {
-        if (subscriber instanceof CleanupSubscriber) {
-            throw new IllegalArgumentException(
-                "CleanupSubscribers must be unregistered with unregisterCleanupSubscriber().");
-        }
-
-        mWrapped.unregister(subscriber);
-
-        synchronized (mSubscribersLock) {
-            if (mSubscribers.remove(subscriber)
-                && mCleanupSubscriber != null
-                && mSubscribers.size() == 0) {
-                mCleanupSubscriber.onAllUnregistered();
-            }
-        }
+        mBus.unregister(subscriber);
     }
 
     @Override public void post(Object event) {
-        mWrapped.post(event);
+        mBus.post(event);
     }
 
-    @Override public void registerCleanupSubscriber(CleanupSubscriber subscriber) {
-        synchronized (mSubscribersLock) {
-            if (mCleanupSubscriber != null) {
-                mWrapped.unregister(subscriber);
-            }
-
-            mCleanupSubscriber = subscriber;
-            mWrapped.register(mCleanupSubscriber);
+    @SuppressWarnings("unused") // Called by reflection from event bus.
+    public void onEvent(NoSubscriberEvent event) {
+        if (event.originalEvent instanceof TypedCursorFetchedEvent<?>) {
+            // If no subscribers were registered for a DataFetchedEvent, then the TypedCursor in
+            // the event won't be managed by anyone else; therefore, we close it ourselves.
+            ((TypedCursorFetchedEvent<?>) event.originalEvent).cursor.close();
+        } else if (event.originalEvent instanceof AppLocationTreeFetchedEvent) {
+            ((AppLocationTreeFetchedEvent) event.originalEvent).tree.close();
         }
-    }
-
-    @Override public void unregisterCleanupSubscriber(CleanupSubscriber subscriber) {
-        // The registered CleanupSubscriber may call this method; however, Java synchronized blocks
-        // are reentrant so synchronizing again is okay.
-        synchronized (mSubscribersLock) {
-            if (mCleanupSubscriber != subscriber) {
-                throw new IllegalStateException(
-                    "A CleanupSubscriber must be registered with registerCleanupSubscriber() "
-                        + "before it can be unregistered.");
-            }
-
-            mWrapped.unregister(subscriber);
-            mCleanupSubscriber = null;
-        }
-    }
-
-    /** Creates a new {@link DefaultCrudEventBus} that wraps the specified {@link EventBus}. */
-    DefaultCrudEventBus(EventBus wrapped) {
-        mWrapped = wrapped;
-        mSubscribers = new HashSet<>();
     }
 }
