+++ app/src/main/java/org/projectbuendia/client/ui/BaseActivity.java
@@ -55,7 +55,7 @@ import de.greenrobot.event.EventBus;
 public abstract class BaseActivity extends FragmentActivity {
     private static final Logger LOG = Logger.create();
     private static final double PHI = (Math.sqrt(5) + 1)/2; // golden ratio
-    private static final double STEP_FACTOR = Math.sqrt(PHI); // each step up/down scales this much
+    private static final double STEP_FACTOR = Math.pow(PHI, 0.25); // each step up/down scales this much
     private static final long MIN_STEP = -2;
     private static final long MAX_STEP = 2;
